Scaffolding to calculate awards
