+++ gamification/models.py
@@ -1,5 +1,3 @@
-from datetime import datetime
-
 from django.contrib.auth import get_user_model
 from django.db import models
 
@@ -35,15 +33,10 @@ class Award(models.Model):
     def __str__(self):
         return f"{self.name}"
 
-    def is_eligible(user: User, award_date: datetime.date):
-        # Check if award is already given
-
-        return True
-
 
 class AwardedAward(models.Model):
     award = models.ForeignKey(Award, on_delete=models.CASCADE)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
 
     def __str__(self):
-        return f"{self.award} awarded to {self.user} in month of {self.award_date.strftime('%b %Y')}"
+        return f"{self.award} awarded to {self.user}"
