+++ pyproject.toml
@@ -18,6 +18,7 @@ python-dateutil = "^2.8.2"
 
 [tool.poetry.dev-dependencies]
 pre-commit = "^2.20.0"
+pylint = "^2.15.2"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
